+++ src/main/java/com/helisa/docmanager/dto/response/SolicitudDetalleResponse.java
@@ -9,11 +9,11 @@ import java.util.List;
 @Data
 @Builder
 public class SolicitudDetalleResponse {
-    private Long id;
+    private Integer id;
     private String estado;
-    private Long tipologiaId;
+    private Integer tipologiaId;
     private LocalDateTime createdAt;
-    private Long createdBy;
+    private Integer createdBy;
     private Boolean ordenFirma;
     private Integer destinatariosTotal;
     private Integer destinatariosAprobados;
